Add Sphinx (docs) skeleton
One of the crucial parts of any software is documentation, because
otherwise no one would know how to use it properly or what problems it
solves. Picobox is not an exception here. So this commit adds a simple
documentation skeleton, in particular:

 * .travis.yml setup
 * tox.ini setup
 * docs/ directory

The documentation is intended to be published on ReadTheDocs, and
corresponding webhook is already installed.