Set proper Git root for use_scm_version battery
When use_scm_version=True current working directory is assumed as Git
root. This is completely wrong assumption that breaks the following
lines:

 $ cd /tmp
 $ python ~/path/to/picobox/setup.py sdist

which might be a valid case on some CI systems. Instead of relying on
CWD pointed to Git root, let's explicitly say that Git root is where
setup.py located.