Use RLock to enable injection on injection case
Injection on injection case is the case when trying to resolve
dependency A, we need to resolve dependency B (or when A depends on B).
In this case we are likely to recursively call .get() method from the
same thread and there's nothing wrong with it and locking must be
relaxed.