Add ChainBox container
ChainBox for boxes is essentially the same as ChainMap for mappings. It
mimics Box interface and hence can substitute one while provides a way
to lookup for a dependency in underlying boxes. While being useful on
its own, ChainBox is essential part of chaining boxes on the stack which
is not implemented yet but will be done pretty soon.