Add chain argument to picobox.push(...)
Picobox stacked interface has one significant drawback so far: it
doesn't allow to override dependencies partially. That means you can
change a box (i.e. DI container) at the top but a new box is required
to provide the very same dependencies.

While this is enough to do basic management it's a little bit
inconvenient in some edge cases like tests. Sometime you need to only
partially override dependencies and there were no way to do so.

So this patch is about introducing 'chain' argument in 'picobox.push()'
interface which basically means "when a dependency is missed, please
lookup for it down the stack".