Rename push() into pushpop()
The reason behind this awkward renaming is that it became clear that we
may need a way to push a box on stack in one place and pop it in
another. This idea basically means we need two standalone functions:
push() and pop(), and therefore the context manager needs to be renamed
in order to free a good name (i.e. push()) for other purpose.

Despite this renaming, picobox follows semantic versioning hence push()
will work as before till next major release - 2.0.