Make it possible to use push() as a regular function
While it's great, that push() works as a context manager and
automatically removes the box from the top of the stack on exit,
it may be useful for push() to behave like a regular function sometimes,
e.g. for integration with frameworks, that have their own notion of
a context stack and want to be able to push/pop boxes explicitly.

pop() is added, so that it's possible to remove the box from the
top of the stack.

While we are at it, also fix usage of threading.Lock in stack.py:
previously a new lock was created in push(), which did not make sense,
as it would effectively mean that concurrent threads would be acquiring
different locks.