Ensure any hashable keys & any values are allowed
Picobox claims to support any hashable keys to identify dependencies,
and any values as dependencies. This property directly depends on
scopes' implementations, so we need to make sure that built-in ones
support this claim. It turns out that `picobox.threadlocal` scope did
not support that, so this patch fixes the issue along with introducing
tests to prevent any regressions in the future.