Add Python 2.7 support
According to Status of Python Branches [1], Python 2.7 is the only
supported python from 2.x era, and it will be supported till 2020. That
said, there are plenty of projects still running on Python 2.7 due to
technical reasons (or legacy) and it would be nice to provide decent DI
for such projects. :)

[1]: https://devguide.python.org/#status-of-python-branches