Isolate dependencies between Flask scopes
There's a bug in Flask scopes where dependencies for the same Flask
scope are shared between different boxes. It turns out that the problem
lies in the fact that the scopes use shared Flask "storages" like "g"
and "current_app". In order to implement isolation, we need to tie
dependencies to a scope instance somehow.