Use 'ruff format', instead of 'black'
Despite being the de facto standard in the community, let's switch from
the black formatter to the ruff formatter. The latter is mostly
compatible with the former but dozen times faster (not relevant here)
and is designed to be compatible with its linting rules. So why use many
tools if I can use just one?

Since picobox is a pet project and is a great playground for me, let's
see how far 'ruff format' can get.