Require either 'value' or 'factory' in Box.put()
Turns out it was possible to call `box.put("key")` without both `value`
and `factory`. Such invocation makes no sense and leads to runtime
errors later on. This patch makes sure that a `TypeError` exception is
raised when `Box.put()` is invoked improperly.

As a minor bonus, this patch updates error messages a bit to better
communicate an error that's occurred.