Keep an async function marker when @pass_ is used
Unfortunately, when the `@picobox.pass_()` decorator is used, a wrapped
coroutine function (i.e. an async function) loses its a coroutine
function marker, i.e. `inspect.iscoroutinefunction()` returns `False`
for such function.

Turns out that there are a lot of software out there that support both
sync and async interfaces, and may choose one based on the type of a
passed function. For instance, Starlette, a web-framework, checks a
provided route function for being a coroutine function before choosing
how to execute in (i.e. in an event loop or in a thread pool).

This patch fixes `@picobox.pass_()` to return a coroutine function when
a wrapped function is also a coroutine function.

Fixes #78