Throw RuntimeError from .pop() when stack is empty
The `picobox.pop()` function used to throw `IndexError` when invoked
with an empty stack. This was never an intentional move, and just
happened because we weren't careful enough.

This patch makes the `.pop()` function to throw `RuntimeError` instead
with a good-looking error message saying that there's no boxes on stack,
similar to what `picobox.put()` or `picobox.get()` are doing.

This patch doesn't introduce a major bump in the version despite being a
backward incompatible change. Invoking `.pop()` when stack is empty is a
symptom of a poorly designed application, and should not be happening in
the first place.