Use a client facade in test_wsgiscopes.py
Previously a bare Flask test client has been used in the tests to make a
request and trigger an endpoint execution. This was error prone, since
the wrong route could have been requested due to a silly typo and we
wouldn't even know. The facade is designed to solve this by always
checking that underlying requests succeed.