Add the MyPy type checker
Mypy is a static type checker for Python. Type checkers help ensure that
you’re using variables and functions in your code correctly. It's about
time I give it a try! This commit lays down basic configuration with a
pretty big ignore list. The ignore list is not to stick here, but rather
temporarly disable the rules until they are resolved.