Skip ASGI websocket tests on PyPy
When running on PyPy, the ASGI websocket tests sometimes hang. This is
especially annoying on CI/CD. Most likely the issue lies either in the
Starlette's TestClient implementation, or anyio it delegates to process
the request in a background thread. For a time being, it's better to
skip them for good.